pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["spine_vision*"]

[project]
name = "spine-vision"
version = "0.1.0"
description = "Medical imaging pipeline for lumbar spine MRI analysis and radiological grading prediction"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "kaleido>=1.2.0",
    "loguru>=0.7.3",
    "nnunetv2",
    "numpy<2",
    "opencv-python>=4.6.0.66",
    "openpyxl>=3.1.5",
    "paddleocr==3.3.2",
    "paddlepaddle==3.2.1",
    "paddlepaddle-gpu>=2.6.2",
    "pandas>=2.3.3",
    "pillow>=10.2.0",
    "plotly>=6.5.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.0",
    "rapidfuzz>=3.14.3",
    "rich>=14.2.0",
    "setuptools>=80.9.0",
    "simpleitk>=2.5.3",
    "torch<2.9",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
    "tyro>=0.9.35",
    "unidecode>=1.4.0",
    "vietocr>=0.3.13",
    "xlrd>=2.0.2",
]

[dependency-groups]
dev = [
    "ipython>=9.8.0",
    "pandas-stubs>=2.3.3.251201",
    "plotly-stubs>=0.1.1",
]

[project.scripts]
spine-vision = "spine_vision.cli:cli"

[tool.uv]
package = true

[tool.uv.sources]
nnunetv2 = { git = "https://github.com/MIC-DKFZ/nnUNet" }

[tool.pyright]
include = ["spine_vision"]
pythonVersion = "3.11"
typeCheckingMode = "basic"
